Enhanced CMakeLists.txt to create a proper export
At the moment miniocpp-config.cmake specifies dependencies, but when
it's actually used by a client it requires that client resolves the
dependencies himself - which means that the client would have to use
find_package() to find everything that minio-cpp uses.

This fixes the problem. It creates miniocpp-targets.cmake, which is
a file generated by cmake, and also creates miniocpp-config.cmake,
which adds the required find_package() commands to the cmake and
then includes miniocpp-targets.cmake, which provides the targets.

This change finalizes the integration of minio-cpp with both cmake
and vcpkg.